boards: phytec: Remove Messtechnik GmbH
PHYTEC has multiple offices in different locations like
the US, France, India and China. The headquarter is located in
Germany. Since now 50% of the PHYTEC boards are from the US
office, we should drop 'Messtechnik GmbH', which is the offical
title of the German office, and keep the name more general.

Signed-off-by: Daniel Schultz <d.schultz@phytec.de>