dts: nxp: Add comment reminders to some DT
The RTXXX and RW61X DT are using syscon compatible
depsite not having a syscon. This is a technical debt
to remain aware of. The reason they use these compatibles
is to use the syscon driver which is a shim to an SDK API
that is somewhat similar to syscon.

Signed-off-by: Declan Snyder <declan.snyder@nxp.com>